Split Anthropic client tests
